refactor(frontend): migrate root layout into features
