+++ src/frontend/src/features/layout/index.ts
@@ -1,0 +1,1 @@
+export { RootHydrateFallback, RootLayout } from "./root-layout";
