+++ src/frontend/src/routes/app.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { RootLayout, RootHydrateFallback } from "@/screens/shell/app-shell-screen";
+import { RootLayout, RootHydrateFallback } from "@/features/layout";
 import { RouteErrorScreen } from "@/routes/-route-error-screen";
 
 export const Route = createFileRoute("/app")({
